Converted common enum raw type to Int16 to match CoreData expectations.

+++ Sources/XcodeServerCommon/CleanSchedule.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 // TODO: Verify Order
-public enum CleanSchedule: Int, Codable {
+public enum CleanSchedule: Int16, Codable {
     case never = 0
     case always = 1
     case daily = 2
