+++ Sources/XcodeServerCommon/EmailType.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// Type of email being sent.
-public enum EmailType: Int, Codable {
+public enum EmailType: Int16, Codable {
     case integrationReport = 0
     case dailyReport = 1
     case weeklyReport = 2
